Don't kill snippets for text property modification
cc-mode modifies text properties in the buffer to cache parsing results,
therefore it's important that yasnippet not kill yasnippets when parsing
is triggered by a function that only wants to parse the
buffer (e.g. which-function-mode).

* yasnippet.el (yas--buffer-pre-modification): New variable.
(yas--on-protection-overlay-modification): Save buffer text (without
properties) on the before change call, and compare on the after change
call.