Remove the key bindings of overridden templates correctly
When loading a second template with the same uuid, the previous template
must be removed from the database including its key binding.  However,
removing the previous template from namehashes and then from keyhash
seems not working, leaving it intact and still expandable.

This fixes the current implementation of `yas--remove-template-by-uuid`
where the uuid equality is tested with `eq` instead of `equal`.