Add ordering to regexp-snippets
Different regexp keys could match the same text(say "[a-z]" and "[a-c]")
while normal keys cannot(I think). This means that an ordering or
priority between snippets become meaningful. For the above example it
would not be clear which key would match the text "abc". If however,
"[a-c]" has higher priority than "[a-z]" there would be only one choice:
"[a-c]" would match.