Prevent infinite recursion when displaying a warning.
If the *Warnings* buffer has auto-fill enabled and we're about to print the
warning, `display-warning' would loop back here through `newline' and
`self-insert-command', causing infinite recursion.

I haven't been able to reproduce this in a noninteractive unit test,
unfortunately, but the issue happens in interactive mode whenever
`yas--original-auto-fill-function' is nil and `auto-fill-mode' is enabled in
the *Warnings* buffer.

* yasnippet.el (yas--auto-fill): Disable `auto-fill-mode' in the *Warnings*
buffer as well.